Compile OnDevicePersonalization against framework-ondevicepersonalization.impl
OnDevicePersonalization uses private apis of
framework-ondevicepersonalization. It appears that using just
"framework-ondevicepersonalization" will compile against either
the stubs or source of framework-ondevicepersonalization depending on if
the module was included in the same apex as it or not.

aosp/2462194 is adding the ability to compile modules outside of the
context of the apex, at which point it will start compiling against
stubs if we don't explicetly specify to use the implementation library.

Bug: 254205429
Test: m javac-check with aosp/2462194
Change-Id: I7debfe3a957e154d4dc486a41bf2848b2b4fa383