Inline literal string value of Flags in API signature files
Previously, due to a mismatch between Metalava and the output of
`java_aconfig_library`, flag constants did not have their literal value
inlined in the API signature files. That would cause problems with
future changes to filter APIs by flags. This change is part of a topic
that updates Metalava to inline the values into `@FlaggedApi` and
update the API signature files.

Bug: 314319629
Test: m checkapi
Change-Id: I1426acdb9dc113d36e233921e0189b19a07b736e