Remove @IntRange from error codes.
The actual range of error codes that can be aggregated is determined by
feedback from privacy review and may be variable depending on the number
of devices. Removing the hardcoded range from the API to prevent
confusion and avoid setting an expectation that all 127 error codes
are available. We can also expand the range in the future if desirable.

Bug: 357914394
Flag: EXEMPT (bug 337358613)
Test: TH
Change-Id: I1da9cb1579b75b0620ebe4a11022f34b72fbcb7b