Verification for ODP sandbox process calls
Updating the logic to retrieve the app caller name when getting the package name, regardless of whether the caller is a sandbox or not. Including sandbox CTS tests to verify the package name is the expected package name regardless of where the call originates from.

Test: atest CtsSandboxOnDevicePersonalizationManagerTests OnDevicePersonalizationManagingServicesTests
Bug: 242792629
Bug: 355209447
Flag: EXEMPT (bug 242792629)
PS_FLAG: BYPASS_PS_FLAG_CHECK

Change-Id: I65c8e4f02f8cb1a0e2667ee13b7f361d02c3ebb8