Convert uses of @hide to block tag
In most places the `@hide` doctag is treated as a block tag on its own
line. However, there are a few places where it is treated as an inline
doctag, e,g. `{@hide}`, and some places where it looks like a block
tag but is inline in the text, e.g. `Some text. @hide`. Dealing with
all the types increases the complexity of Metalava and makes it
slightly slower.

This change replaces the non-block tag forms with a block tag.

BYPASS_PS_FLAG_CHECK: minor build change that does not change API or behavior

Flag: EXEMPT PURE_REFACTOR
      making @hide a block tag does not the change API or behavior
Bug: 429965593
Test: m checkapi
Change-Id: I0e1fd6071e3e1100bdbad06a8b82c7cb25b757f0